@@ -1,55 +1,56 @@
 {
   "name": "@studio-freight/hamo",
-  "version": "0.4.32",
+  "version": "0.4.33",
   "description": "hamo means hook, do the math.",
   "author": "studio-freight",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/studio-freight/hamo.git"
-  },
-  "type": "module",
-  "sideEffects": false,
-  "source": "src/index.js",
-  "module": "dist/hamo.esm.js",
+  "main": "dist/index.js",
   "scripts": {
     "dev": "vite docs",
-    "build": "microbundle src/index.js",
-    "prepublishOnly": "npm run build"
+    "size": "size-limit",
+    "clean": "rimraf dist",
+    "build": "tsc",
+    "prebuild": "npm run clean",
+    "preversion": "npm run build",
+    "postversion": "git push --follow-tags"
   },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/studio-freight/bibliotheca.git"
+  },
+  "keywords": [],
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/studio-freight/bibliotheca/issues"
+  },
+  "homepage": "https://github.com/studio-freight/bibliotheca#readme",
   "devDependencies": {
-    "@vitejs/plugin-react": "^1.3.2",
-    "eslint": "8.15.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react": "^7.30.0",
-    "microbundle": "^0.15.0",
-    "path": "^0.12.7",
-    "prettier": "^2.6.2",
-    "prettier-eslint": "^15.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "vite": "^2.9.9"
-  },
-  "dependencies": {
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "@studio-freight/tempus": "^0.0.25",
+    "@size-limit/preset-small-lib": "^8.0.1",
     "nanoid": "^4.0.0",
-    "react-use": "^17.3.2",
-    "throttle-debounce": "^5.0.0"
+    "path": "^0.12.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rimraf": "^3.0.2",
+    "size-limit": "^8.0.1",
+    "throttle-debounce": "^5.0.0",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.9",
+    "@studio-freight/tempus": "^0.0.25"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "nanoid": "^4.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "throttle-debounce": "^5.0.0",
+    "@studio-freight/tempus": "^0.0.25"
   },
-  "bugs": {
-    "url": "https://github.com/studio-freight/hamo/issues"
-  },
-  "homepage": "https://github.com/studio-freight/hamo#readme",
-  "directories": {
-    "doc": "docs"
-  },
-  "keywords": [
-    "hooks",
-    "react"
+  "size-limit": [
+    {
+      "limit": "2 kB",
+      "path": "dist/index.js"
+    }
   ]
 }
